% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_categories.R
\name{aggregate_categories}
\alias{aggregate_categories}
\title{Aggregation for Oakland variables.}
\usage{
aggregate_categories(dat, compute = "mean", group_vars = c(NULL))
}
\arguments{
\item{dat}{Data with a column containing census tracts ("tractid10") and categorical variables to group by or numeric variables to aggregate.}

\item{compute}{"mean", "median", or "sum"}

\item{group_vars}{Optional list of variables to group_by before aggregation. Note: All non-numeric variables should either be included here or de-selected.}
}
\value{
Aggregated data frame.
}
\description{
Take mean, median, or sum of all variables by ethnoracial, income, and gentrification grouping.
}
