% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landlord_location_stacked_bar.R
\name{landlord_location_stacked_bar}
\alias{landlord_location_stacked_bar}
\title{Produce stacked bar chart panel of landlord location by neighborhood category.}
\usage{
landlord_location_stacked_bar(
  dat,
  save = F,
  savename = "plot.png",
  caption = paste0(frb_acs_caption, ses_caption, period_caption)
)
}
\arguments{
\item{dat}{oak-data-cleaning/County_Assessor/landlords_clean.csv}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Caption for figure}
}
\value{
Stacked bar chart panel of landlord location by neighborhood category.
}
\description{
This function takes in data and produces a stacked bar chart panel of landlord
location grouped by ethnoracial, income, gentrification,
}
