% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_graph.R
\name{line_graph}
\alias{line_graph}
\title{Produce line graph of variable by periods}
\usage{
line_graph(
  dat,
  var,
  limits,
  title = NULL,
  y_title = NULL,
  save = F,
  savename = "plot.png",
  caption = ""
)
}
\arguments{
\item{dat}{Data with a columns containing variable of interest, character periods variable ("periods"), and grouping variable ("cat").}

\item{var}{Name of variable to plot.}

\item{y_title}{Title to display along y-axis.}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Caption for figure}
}
\value{
Line graph of variable by periods, grouped by race, ethnoracial, income, or gentrification category.
}
\description{
This function takes in data and produces a line graph of variables by periods,
grouped by neighborhood category or race.
}
