% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_discrete_map.R
\name{make_discrete_map}
\alias{make_discrete_map}
\title{Produce Oakland map of discrete variable}
\usage{
make_discrete_map(
  shp_tracts,
  discrete_cat = "gent",
  coord = F,
  save = F,
  savename = "plot.png",
  caption = paste0(frb_caption, ses_caption, period_caption)
)
}
\arguments{
\item{shp_tracts}{"US_tract_2010.shp" loaded object}

\item{discrete_cat}{Discrete category to plot: gent, income, ethnoracial}

\item{coord}{T if plotting coordinate values (lat, lon). Default is F.}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Figure caption}
}
\value{
Map of discrete variable: gentrification, income, or ethnoracial category.
}
\description{
This function takes in category (gentrification, income, or ethnoracial) and
produces a census tract map of Oakland representing a discrete variable.
}
