% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Produce King County map of continuous variable with gradient color scale}
\usage{
make_map(
  data,
  var,
  shp_file,
  palette = "sequential",
  jenksbreaks = T,
  neg_bins = 3,
  pos_bins = 3,
  breaks = scales::extended_breaks(n = 7),
  labels = scales::percent,
  limits = NULL,
  coord = F,
  save = F,
  savename = "plot.png",
  title = NULL,
  caption = "test"
)
}
\arguments{
\item{data}{Data with a column containing HRAs and variable of interest.}

\item{var}{Name of column containing variable to plot.}

\item{shp_file}{Hra shape file}

\item{palette}{Color palette: "sequential" (default) or "diverging"}

\item{jenksbreaks}{Uses Jenks breaks when T, otherwise uses continuous color scale}

\item{neg_bins}{For Jenks breaks, number of negative color bins. Default is 3.}

\item{pos_bins}{For Jenks breaks, number of positive color bins. Default is 3.}

\item{breaks}{Gradient scale breaks, either numeric vector or scales::extended_breaks(n = 6)}

\item{labels}{Gradient scale labels, either character vector or scales::percent or scales::comma}

\item{limits}{Manual limits for color scale, numeic vector: c(min, max)}

\item{coord}{T if plotting coordinate values (lat, lon). Default is F.}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{title}{Title of figure}

\item{caption}{Figure caption}
}
\value{
Map of variable of interest.
}
\description{
This function takes in data and produces an HRA map of King County
representing the variable using a gradient color scale. HRA
column must be named "HRA".
}
