% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map_panel.R
\name{make_map_panel}
\alias{make_map_panel}
\title{Produce King County map panel of continuous variable with gradient color scale}
\usage{
make_map_panel(
  data,
  var,
  shp_file,
  palette = "sequential",
  jenksbreaks = T,
  neg_bins = 3,
  pos_bins = 3,
  breaks = scales::extended_breaks(n = 6),
  labels = scales::percent,
  limits = NULL,
  coord = F,
  save = F,
  savename = "plot.png",
  caption = paste0(frb_caption, ses_caption, period_caption)
)
}
\arguments{
\item{data}{Data with a column containing HRAs; 2, 3, or 4 distinct periods; and variable of interest.}

\item{var}{Name of column containing variable to plot.}

\item{shp_file}{hra shapefiles loaded object}

\item{palette}{Color palette: "sequential" (default) or "diverging"}

\item{jenksbreaks}{Uses Jenks Breaks when T, otherwise uses continuous color scale}

\item{neg_bins}{For Jenks breaks, number of negative color bins. Default is 3.}

\item{pos_bins}{For Jenks breaks, number of positive color bins. Default is 3.}

\item{breaks}{Gradient scale breaks, either numeric vector or scales::extended_breaks(n = 6)}

\item{labels}{Gradient scale labels, either character vector or scales::percent or scales::comma}

\item{limits}{Gradient scale limits, c(min, max)}

\item{coord}{T if plotting coordinate values (lat, lon). Default is F.}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Figure caption}
}
\value{
Map panel of variable of interest across 2, 3, or 4 periods.
}
\description{
This function takes in data and produces a panel of HRA maps of
King County representing the variable using a gradient or discrete color scale, across
distinct periods. Should have HRA column and "periods" column for distinct maps.
}
