% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_periods_ses.R
\name{plot_bar_periods_ses_grid}
\alias{plot_bar_periods_ses_grid}
\title{Produce grouped bar chart panel across four periods and SES category.}
\usage{
plot_bar_periods_ses_grid(
  dat,
  var,
  limits,
  y_title = NULL,
  group = "gent",
  save = F,
  savename = "plot.png",
  caption = paste0(frb_acs_caption, ses_caption, period_caption)
)
}
\arguments{
\item{dat}{Data with a column containing census tracts and variable of interest. Aggregated data with column "period"}

\item{var}{Name of column containing variable to plot.}

\item{limits}{Y-axis limits}

\item{y_title}{Title to display along y-axis.}

\item{group}{Category for x-axis grouping: "gent" (default), "ethnoracial", "income", "ses", or "period"}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Caption for figure}
}
\value{
Grouped bar chart panel across four periods and SES categories.
}
\description{
This function takes in data and produces a bar chart panel grouped by either
gentrification, ethnoracial, income, ses, or period category across four
housing periods and SES categories.
}
