% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_ses.R
\name{plot_bar_ses}
\alias{plot_bar_ses}
\title{Produce stacked bar charts with SES facets and period x-axis grouping.}
\usage{
plot_bar_ses(
  dat,
  fill = "dest",
  y_title = NULL,
  save = F,
  savename = "plot.png",
  caption = paste0(frb_caption, ses_caption, period_caption)
)
}
\arguments{
\item{dat}{Data with a column containing variables of interest and grouping variables.}

\item{fill}{Name of column containing fill variable: "move", "dest" (default).}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Caption for figure}
}
\value{
Stacked bar charts with SES facets, period x-axis groupings
}
\description{
This function takes in data and produces a panel of stacked bar charts,
based on different groupings. Intended to be used with oakViz::aggregate_categories()
}
