% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lollipop.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Produce lollipop plot by King County HRA.}
\usage{
plot_lollipop(
  data,
  var,
  limits,
  title = NULL,
  x_title = NULL,
  scale_type = "numeric",
  save = F,
  savename = "plot.png",
  caption = paste0(frb_acs_caption_splitline, ses_caption)
)
}
\arguments{
\item{data}{Data with column for variable of interest with "facet" and "facet_col"}

\item{var}{Column name of variable of interest.}

\item{limits}{Y-axis limits.}

\item{title}{Plot title}

\item{x_title}{Title to display along x-axis}

\item{scale_type}{Y-axis scale type: "numeric" or "percent"}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Figure caption}
}
\value{
Lollipop plot of variable by HRA and SES.
}
\description{
This function takes in data and produces a horizontal lollipop plot by
King County HRA The order of these categories can be adjusted by changing
the factor levels of the facet variable. Input data needs columns for
variable of interest (titled "var") and HRA.
}
