% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses_period_map_panel.R
\name{ses_period_map_panel}
\alias{ses_period_map_panel}
\title{Produce 4x4 map panel of variable of interest in Oakland}
\usage{
ses_period_map_panel(
  data,
  var,
  shp_tracts,
  palette = "sequential",
  jenksbreaks = T,
  neg_bins = 3,
  pos_bins = 3,
  breaks = scales::extended_breaks(n = 6),
  labels = scales::comma,
  limits = NULL,
  save = F,
  savename = "plot.png",
  caption = paste0(frb_caption, ses_caption, period_caption)
)
}
\arguments{
\item{data}{Data with a column containing tractid10, year, ses, and variable of interest.}

\item{var}{Name of column containing variable to plot (absolute count, not pct).}

\item{shp_tracts}{"US_tract_2010.shp" loaded object}

\item{palette}{Color palette: "sequential" (default) or "diverging"}

\item{jenksbreaks}{Uses Jenks Breaks when T, otherwise uses continuous color scale}

\item{neg_bins}{For Jenks breaks, number of negative color bins. Default is 3.}

\item{pos_bins}{For Jenks breaks, number of positive color bins. Default is 3.}

\item{breaks}{Gradient scale breaks, either numeric vector or scales::extended_breaks(n = 6)}

\item{labels}{Gradient scale labels, either character vector or scales::percent or scales::comma}

\item{limits}{Gradient scale limits, c(min, max)}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Figure caption}
}
\value{
4x4 panel maps of variable of interest, by SES and period.
}
\description{
This function takes in data and produces a 4x4 panel of data across SES and periods.
Not intended to be used with aggregated CCP data.
Assumes 5\% sample of population, all counts are multiplied by 20.
}
