% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_bar.R
\name{stacked_bar}
\alias{stacked_bar}
\title{Produce panel of stacked bar charts}
\usage{
stacked_bar(
  dat,
  fill = "ses",
  group = "period",
  y_title = NULL,
  save = F,
  savename = "plot.png",
  caption = paste0(frb_acs_caption, ses_caption, period_caption)
)
}
\arguments{
\item{dat}{Data with a column containing variables of interest and grouping variables.}

\item{fill}{Name of column containing fill variable: "ses" (default), "move", "dest".}

\item{group}{Category for x-axis grouping: "period" (default), "ses"}

\item{y_title}{Title to display along y-axis.}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Caption for figure}
}
\value{
Panel of stacked bar charts of different groupings
}
\description{
This function takes in data and produces a panel of stacked bar charts,
based on different groupings. Intended to be used with oakViz::aggregate_categories()
}
