% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_bar_ses.R
\name{stacked_bar_ses}
\alias{stacked_bar_ses}
\title{Produce panel of stacked bar charts of SES by a 'periods' column}
\usage{
stacked_bar_ses(
  dat,
  var,
  y_title = NULL,
  save = F,
  savename = "plot.png",
  title = NULL,
  caption = "test"
)
}
\arguments{
\item{dat}{Data with a column containing variables of interest and grouping variables.}

\item{var}{Name of column containing variable to plot}

\item{y_title}{Title to display along y-axis.}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Caption for figure}
}
\value{
Panel of stacked bar charts of different groupings
}
\description{
This function takes in data and produces a panel of stacked bar charts,
based on different groupings. Intended to be used with oakViz::aggregate_categories()
}
