% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map_panel.R
\name{make_map_panel}
\alias{make_map_panel}
\title{Produce Oakland map panel of continuous variable with gradient color scale}
\usage{
make_map_panel(
  data,
  var,
  shp_tracts,
  palette = "sequential",
  jenksbreaks = T,
  neg_bins = 3,
  pos_bins = 3,
  breaks = scales::extended_breaks(n = 6),
  labels = scales::percent,
  limits = NULL,
  coord = F,
  save = F,
  savename = "plot.png",
  caption = paste0(frb_caption, ses_caption, period_caption)
)
}
\arguments{
\item{data}{Data with a column containing census tracts, distinct periods, and variable of interest.}

\item{var}{Name of column containing variable to plot.}

\item{shp_tracts}{"US_tract_2010.shp" loaded object}

\item{palette}{Color palette: "sequential" (default) or "diverging"}

\item{jenksbreaks}{Uses Jenks Breaks when T, otherwise uses continuous color scale}

\item{neg_bins}{For Jenks breaks, number of negative color bins. Default is 3.}

\item{pos_bins}{For Jenks breaks, number of positive color bins. Default is 3.}

\item{breaks}{Gradient scale breaks, either numeric vector or scales::extended_breaks(n = 6)}

\item{labels}{Gradient scale labels, either character vector or scales::percent or scales::comma}

\item{limits}{Gradient scale limits, c(min, max)}

\item{coord}{T if plotting coordinate values (lat, lon). Default is F.}

\item{save}{T if user would like to return plot object and save file, F (default) to just return object.}

\item{savename}{File name of map for saving.}

\item{caption}{Figure caption}
}
\value{
Map panel of variable of interest across four periods.
}
\description{
This function takes in data and produces a panel of census tract maps of
Oakland representing the variable using a gradient color scale, across
four distinct periods. Should have "tractid10" column for census tracts
and "periods" column for distinct time periods.
}
